requirements.txt
gradio>=4.0.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.39.0
google-generativeai>=0.8.0
